Removed AnchoredOperation.
