+++ tests/core/DocumentComposer.spec.ts
@@ -2,7 +2,6 @@ import DocumentComposer from '../../lib/core/versions/latest/DocumentComposer';
 import DocumentModel from '../../lib/core/versions/latest/models/DocumentModel';
 import ErrorCode from '../../lib/core/versions/latest/ErrorCode';
 import KeyUsage from '../../lib/core/versions/latest/KeyUsage';
-import Operation from '../../lib/core/versions/latest/Operation';
 import SidetreeError from '../../lib/core/SidetreeError';
 
 describe('DocumentComposer', async () => {
@@ -164,7 +163,7 @@ describe('DocumentComposer', async () => {
         }
       ];
 
-      Operation.applyPatchesToDidDocument(didDocument, patches);
+      (DocumentComposer as any).applyPatchesToDidDocument(didDocument, patches);
 
       expect(didDocument.publicKey).toEqual([
         { id: 'aRepeatingId', type: 'someType', usage: 'some usage', controller: 'someId' },
